further spcec refactoring
